% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelCompare.R
\name{compareModels}
\alias{compareModels}
\title{compute approximate Bayes Factor in favor of one spatial SEIR model over another}
\usage{
compareModels(modelList, priors = NA, n_samples = 1000,
  batch_size = 10000, max_itrs = 1000)
}
\arguments{
\item{modelList}{A list of models to compare with approximate Bayes factors.
function.}

\item{priors}{The prior probabilities of each model in \code{modelList}}

\item{n_samples}{The desired number of accepted simulation values on which the
Bayes Factor calculation is to be based}

\item{batch_size}{The number of epidemics to simulate in parallel before
assessing the number of accepted samples}

\item{max_itrs}{The maximum number of parallel batches to execute before giving up}
}
\description{
compute approximate Bayes Factor in favor of one spatial SEIR model over another
}
\details{
A Bayes Factor is a measure of the posterior evidence in favor
of one model compared to another. In the ABC setting, we may compute
approximate Bayes Factors of comparably converged models by assessing
the parameter acceptance rate at a new iteration.
}
\examples{
\dontrun{compareModels(list(model1, model2))}
}

