% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataModel.R
\name{DataModel}
\alias{DataModel}
\title{Create a DataModel object, describing the relationship of your data to the SEIR model}
\usage{
DataModel(
  Y,
  type = c("identity", "overdispersion"),
  compartment = c("I_star", "R_star", "I"),
  cumulative = FALSE,
  params = list()
)
}
\arguments{
\item{Y}{A matrix with T rows and n columns, for time points and spatial locations respectively.}

\item{type}{A string equal to "identity", "overdispersion", or "fractional"}

\item{compartment}{A string equal to "I_star" or "R_star"}

\item{cumulative}{A logical value indicating whether the data is reported on a cumulative scale}

\item{params}{A list of optional parameters}
}
\value{
an object of type \code{\link{DataModel}}
}
\description{
Create a DataModel object, describing the relationship of your data to the SEIR model
}
\details{
A fundamental task in building hierarchical models is to describe the way in which
 the observed data relates to the underlying model. Here, we assume that the 
 matrix of observed values, \code{Y}, is related to one of two epidemic compartments:
 I_star or R_star. In the notation of Brown, Porter, and Oleson 2015, these correspond
 to the $I^*$ and $R^*$ compartments, which catalog the newly infectious and recovered
 individuals, respectively. For a discussion of these definitions, in addition to a more
 detailed description of the overall spatial SEIR framework, please refer to that work. 
 Depending on the data model type, additional parameters may be required:
 \itemize{
 \item{phi: }{An overdispersion parameter, required for the 'overdispersion' data model type.}
 \item{report_fraction}{The (scalar) estimated reporting fraction for the epidemic. 
 Required for the 'fractional' data model.}
 \item{report_fraction_ess}{The effective sample size associated with the required reporting fraction
  in the 'fractional' data model.}}
}
\examples{
dataModel = DataModel(rpois(100, 10)) 
}
