% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceModel.R
\name{DistanceModel}
\alias{DistanceModel}
\title{Create a DistanceModel object describing the network structure of the 
population under study.}
\usage{
DistanceModel(
  distanceList,
  scaleMode = c("none", "rowscale", "invsqrt"),
  priorAlpha = 1,
  priorBeta = 1
)
}
\arguments{
\item{distanceList}{a list of square, symmetric distance matrices.}

\item{scaleMode}{an optional argument specifying the type of 
preprocessing needed.}

\item{priorAlpha}{the first shape parameter for the beta distributed
autocorrelation terms}

\item{priorBeta}{the second shape parameter for the beta distributed
autocorrelation terms}
}
\value{
an object of type \code{\link{DistanceModel}}
}
\description{
Create a DistanceModel object describing the network structure of the 
population under study.
}
\details{
In stochastic spatial SEIR models as specified in Brown et al. 2015, 
 populations are divided into homogeneous groups, or locations, with
 heterogeneous mixing between groups. This is accomplished 
 using a distance matrix parameterization, in which some number of 
 square, symmetric distance matrices are constructed, each of which 
 receives a spatial autocorrelation parameter. 

 Care must be taken to specify reasonable prior parameters for such
 terms, as well as in the construction and scaling of the distance 
 matrices; it is certainly possible to construct an overspecified
 model, and to correspondingly bias inference about other important
 exposure process terms.
}
\examples{
distanceModel <- DistanceModel(list(1-diag(4)))
}
