% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposureModel.R
\name{ExposureModel}
\alias{ExposureModel}
\title{Create an ExposureModel object, describing an epidemic intensity process}
\usage{
ExposureModel(X, nTpt, nLoc, betaPriorPrecision = NA, betaPriorMean = NA,
  offset = NA)
}
\arguments{
\item{X}{an $(n*T)$ by $p$ design matrix, where $n$ is the number
of locations, $T$ is the number of time points, and $p$ is the number
of exposure process parameters. Each column corresponds to a parameter,
while each block of $T$ rows corresponds to the time series of covariate
values associated with a location.}

\item{nTpt}{the number of time points, $T$}

\item{nLoc}{the number of locations, $N$}

\item{betaPriorPrecision}{the prior precisions of the $p$ exposure 
process parameters}

\item{betaPriorMean}{the prior means of the $p$ exposure process
parameters}

\item{offset}{a vector of $T$ temporal offset terms, capturing the 
relative ammount of aggregated continuous time corresponding to each
recorded discrete time point.}
}
\description{
Create an ExposureModel object, describing an epidemic intensity process
}
\details{
The exposure process allows the inclusion of both spatially and temporally
varying covariates, as well as invariant quantities (intercepts, demographic 
features etc.).
}
\examples{
exposureModel <- ExposureModel(cbind(1, (1:25)/25),
                                         nTpt = 25, nLoc = 1,
                                         betaPriorPrecision = 0.1,
                                         betaPriorMean = 0)
}

