% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialValueContainer.R
\name{InitialValueContainer}
\alias{InitialValueContainer}
\title{Create an InitialValueContainer object, which catalogues the starting population
values for an epidemic.}
\usage{
InitialValueContainer(
  S0,
  E0,
  I0,
  R0,
  type = c("identity", "uniform"),
  params = list()
)
}
\arguments{
\item{S0}{An integer vector of susceptible counts}

\item{E0}{An integer vector of exposed counts}

\item{I0}{An integer vector of infectious counts}

\item{R0}{AN integer vector of removed counts}
}
\description{
Create an InitialValueContainer object, which catalogues the starting population
values for an epidemic.
}
\details{
While approximate Bayesian techniques avoid the need to specify starting values
 for the spatiotemporally varying population values (i.e. the current number of 
 susceptible, exposed, infectious, and removed individuals), we still need
 to specify how many individuals start out in each of these categories. In the 
 future, we may support the specification of priors on these terms, but for
 now they are considered fixed.
}
\examples{
inits <- InitialValueContainer(c(1000,1000),
                                          c(10,0),
                                          c(0,0),
                                          c(100,100))
}
