% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Kikwit1995}
\alias{Kikwit1995}
\title{Ebola Case Counts by Day - 1995 Outbreak in Kikwit}
\format{
A data frame with 197 rows and 2 variables
\describe{
    \item{Date}{case index date}
    \item{Count}{number of cases observed}
}
}
\usage{
Kikwit1995
}
\description{
A dataset containing Ebola case counts by day between 3-6-1995 and 9-18-1995.
}
\references{
Lekone, P.E. and Finkenstadt, B.F. (2006). Statistical inference in a stochastic epidemic
SEIR model with control intervention: Ebola as a case study. Biometrics 62,1170-1177
}
\keyword{datasets}
