% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SpatialSEIRModel.R
\name{plot.SpatialSEIRModel}
\alias{plot.SpatialSEIRModel}
\title{Plot a graphical summary of the marginal posterior distribution of SEIR model parameters}
\usage{
\method{plot}{SpatialSEIRModel}(x, ...)
}
\arguments{
\item{x}{a \code{\link{SpatialSEIRModel}} object}

\item{\dots}{additional arguments to be passed to plotting functions.}
}
\description{
Plot a graphical summary of the marginal posterior distribution of SEIR model parameters
}
\examples{
\dontrun{plot(modelObject)}
}

