% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_summary_SpatialSEIRModel.R
\name{print.summary.SpatialSEIRModel}
\alias{print.summary.SpatialSEIRModel}
\title{Print a \code{summary.SpatialSEIRModel} object}
\usage{
\method{print}{summary.SpatialSEIRModel}(x, ...)
}
\arguments{
\item{x}{a \code{summary.SpatialSEIRModel} object}

\item{\dots}{not used}
}
\description{
Print a \code{summary.SpatialSEIRModel} object
}

