% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinfectionModel.R
\name{ReinfectionModel}
\alias{ReinfectionModel}
\title{Create a ReinfectionModel object, which determines whether and how individuals
may return to the susceptible population after an infection runs its course.}
\usage{
ReinfectionModel(reinfectMode = c("SEIR", "SEIRS", "Fixed"),
  X_prs = NA, priorPrecision = NA, priorMean = NA)
}
\arguments{
\item{reinfectMode}{a string equal to one of: \code{"SEIR"}
, \code{"SEIRS"}, or \code{"Fixed"}, SEIR models include no
reinfection. SEIRS models include a reinfection process,
and Fixed models allow reinfection, but do not estimate the parameters
involved. Currently, only SEIR and SEIRS modes are supported.}

\item{X_prs}{a $T$ by $p$ design matrix, which must be specified for SEIRS and
Fixed models.}

\item{priorPrecision}{the prior precisions for each of the reinfection parameters}

\item{priorMean}{the prior means for each of the reinfection parameters}
}
\description{
Create a ReinfectionModel object, which determines whether and how individuals
may return to the susceptible population after an infection runs its course.
}
\details{
SEIR models are probably the most commonly employed version of compartmental
 epidemic models today, and are specified using a simple SEIR reinfection model
 (i.e., a model which introduces no reinfection process). On the other hand, 
 there are some diseases which must be modeled more flexibly, for individuals
 are not permanently (over the course of a typical study period) removed from
 the susceptible population by immunity or death after an infection.
}
\examples{
reinfectionModel = ReinfectionModel("SEIR")
}
