% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSim.R
\name{Rsim}
\alias{Rsim}
\title{Simulate from a SEIR/SEIRS process using R (debug feature)}
\usage{
Rsim(seed, params, exposure_model, reinfection_model, distance_model,
  transition_priors, initial_value_container)
}
\arguments{
\item{seed}{A random seed}

\item{params}{Parameters needed for the simulation, a list including beta_SE, 
beta_RS, rho, gamma_EI, gamma_IR, and optionally path-specific parameters.}

\item{exposure_model}{An exposure model object, which describes the
spatial and temporal variability of the exposure/infection process. Valid 
exposure models are created using the \code{\link{ExposureModel}} function.}

\item{reinfection_model}{A reinfection model object, which describes
whether or not individuals are able to return from the Removed category
to the Susceptible population. Valid reinfection models are created using the
\code{\link{ReinfectionModel}} function.}

\item{distance_model}{A distance model object which describes the underlying
contact network, in addition to prior parameters which constrain the contact
process. Valid distance models are created using \code{\link{DistanceModel}}}

\item{transition_priors}{An object containing information about the E to I and
I to R transition prior parameters. These are created using the 
\code{\link{TransitionPriors}} function, and involve either transition
probabilities and corresponding effective sample sizes (amount of prior 
information), or manually specified probability distributions. .}

\item{initial_value_container}{An object specifying the initial state
of the epidemic for each spatial location, created by the
\code{\link{InitialValueContainer}} function.}
}
\value{
a list
}
\description{
Simulate from a SEIR/SEIRS process using R (debug feature)
}
\details{
This function is provided for testing and debugging purposes: it's not intended
for the end user
}
\examples{
\dontrun{results = Rsim(seed, 
                                   beta_SE,
                                   beta_RS, 
                                   rho,
                                   exposure_model,
                                   reinfection_model,
                                   distance_model,
                                   transition_priors,
                                   initial_value_container)}
}

