% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingControl.R
\name{SamplingControl}
\alias{SamplingControl}
\title{Create a SamplingControl object, which determines which ABC algorithm is 
to be used, and how it is configured.}
\usage{
SamplingControl(seed, n_cores, algorithm = "Beaumont2009", params = NA)
}
\arguments{
\item{seed}{an integer, giving the seed to be used when simulating epidemics}

\item{n_cores}{an integer giving the number of CPU cores to employ}

\item{algorithm}{a string, either equal to "BasicABC" for the simple
ABC rejection algorithm of Rubin (1980), "Beaumont2009" for the
SMC approach of Beaumont et al. (2009), or "DelMoral2006"  for the adaptive
SMC approach of Del Moral (2012).}

\item{params}{optional algorithm configuration parameters, see: detail.}
}
\value{
an object of type \code{\link{SamplingControl}}
}
\description{
Create a SamplingControl object, which determines which ABC algorithm is 
to be used, and how it is configured.
}
\examples{
samplingControl <- SamplingControl(123123, 2)
}
