% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRModel.R
\name{summary.SpatialSEIRModel}
\alias{summary.SpatialSEIRModel}
\title{Produce a summary of the results of fitting a SEIR model}
\usage{
\method{summary}{SpatialSEIRModel}(object, ...)
}
\arguments{
\item{object}{a \code{\link{SpatialSEIRModel}} object}

\item{\dots}{not used}
}
\value{
a summary.SpatialSEIRModel object
}
\description{
Produce a summary of the results of fitting a SEIR model
}

