% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SEIRModel.R
\name{update.SpatialSEIRModel}
\alias{update.SpatialSEIRModel}
\title{Update a \code{\link{SpatialSEIRModel}} object by drawing additional samples.}
\usage{
\method{update}{SpatialSEIRModel}(object, ...)
}
\arguments{
\item{object}{a \code{\link{SpatialSEIRModel}} object}

\item{\dots}{Additional arguments include \code{sampling_control}, which
must be a \code{\link{SamplingControl}} object, \code{samples} which indicates
the number of parameter values to sample from the posterior, and \code{verbose}
which indicates whether verbose output should be provided.}
}
\description{
Update a \code{\link{SpatialSEIRModel}} object by drawing additional samples.
}

