% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transitionPriors.R
\name{WeibullTransitionPriors}
\alias{WeibullTransitionPriors}
\title{Build a path specific Weibull TransitionPriors object, with
independent gamma priors for the shape/scale parameters of the
latent and infectious time distributions.}
\usage{
WeibullTransitionPriors(latent_shape_prior_alpha, latent_shape_prior_beta,
  latent_scale_prior_alpha, latent_scale_prior_beta,
  infectious_shape_prior_alpha, infectious_shape_prior_beta,
  infectious_scale_prior_alpha, infectious_scale_prior_beta)
}
\arguments{
\item{latent_shape_prior_alpha}{The alpha (shape) value for the
gamma hyperprior on the latent period shape parameter}

\item{latent_shape_prior_beta}{The beta (rate) value for the gamma
hyperprior on the latent period shape parameter}

\item{latent_scale_prior_alpha}{The alpha (shape) value for the gamma
hyperprior on the latent period scale parameter}

\item{latent_scale_prior_beta}{The beta (rate) value for the gamma
hyperprior on the latent period scale parameter}

\item{infectious_shape_prior_alpha}{The alpha (shape) value for the gamma
hyperprior on the infectious period shape parameter}

\item{infectious_shape_prior_beta}{The beta (rate) value for the gamma
hyperprior on the infectious period shape parameter}

\item{infectious_scale_prior_alpha}{The alpha (shape) value for the gamma
hyperprior on the infectious period scale parameter}

\item{infectious_scale_prior_beta}{The beta (rate) value for the gamma
hyperprior on the infectious period scale parameter}
}
\description{
Build a path specific Weibull TransitionPriors object, with
independent gamma priors for the shape/scale parameters of the
latent and infectious time distributions.
}
\details{
WeibullTransitionPriors assumes that both the latent and infectious times
are distributed according to Weibull distributions. Let \eqn{\tau_{lat}} and
\eqn{\tau_{inf}} represent random variables corresponding to latent and infectious
times. We then have:
\eqn{\tau_{lat} \sim Weibull(\theta_1^{lat}, \theta_2^{lat})}
\eqn{\tau_{inf} \sim Weibull(\theta_1^{inf}, theta_2^{inf})}
For each of these four parameters, we assign a gamma prior, parameterized by
shape and rate:
\eqn{f(\theta_{.}^{.}| \alpha, \beta^{-1}) = frac{1}{\Gamma(\alpha) \beta^\alpha}e^{-\frac{x}{\beta}}x^{\alpha-1}}

These hyperparameters are specified by the eight arguments to WeibullTransitionPriors.
}
\examples{
transitionPriors <- WeibullTransitionPriors(1,1,1,1,1,1,1,1)
}
\seealso{
\code{\link{TransitionPriors}}, \code{\link{PathSpecificTransitionPriors}},
\code{\link{ExponentialTransitionPriors}}
}

