% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{WestAfrica2015}
\alias{WestAfrica2015}
\title{Ebola Case Count reports - 2014 and 2015 outbreak in West Africa (CDC/WHO)}
\format{
A data frame with 133 rows and 18 variables
\describe{
    \item{Date}{Case report date}
    \item{Day}{Day index}
    \item{Cases_Guinea}{Cumulative cases - Guinea}
    \item{Cases_Liberia}{Cumulative cases - Liberia}
    \item{Cases_SierraLeone}{Cumulative cases - Sierra Leone}
    \item{Cases_Nigeria}{Cumulative cases - Nigeria}
    \item{Cases_Senegal}{Cumulative cases - Senegal}
    \item{Cases_UnitedStates}{Cumulative cases - United States}
    \item{Cases_Spain}{Cumulative cases - Spain}
    \item{Cases_Mali}{Cumulative cases - Mali}
    \item{Deaths_Guinea}{Cumulative deaths - Guinea }
    \item{Deaths_Liberia}{Cumulative deaths - Liberia }
    \item{Deaths_SierraLeone}{Cumulative deaths - Sierra Leone }
    \item{Deaths_Nigeria}{Cumulative deaths - Nigeria }
    \item{Deaths_Senegal}{Cumulative deaths - Senegal }
    \item{Deaths_UnitedStates}{Cumulative deaths - United States }
    \item{Deaths_Spain}{Cumulative deaths - Spain }
    \item{Deaths_Mali}{Cumulative deaths - Mali }
}
}
\source{
\url{https://github.com/cmrivers/ebola}
}
\usage{
WestAfrica2015
}
\description{
A dataset containing cumulative Ebola case counts between 3-22-2014 and 3-29-2015
}
\keyword{datasets}
