\name{ABSEIR-package}
\alias{ABSEIR-package}
\docType{package}
\title{ABSEIR}
\description{Approximate Bayesian Computation for Stochastic Compartmental Models}
\details{
  Approximate Bayesian Computation is a natural framework within which to fit
  epidemic models, for it is often far easier to simulate epidemics than it is
  to evaluate their posterior distribution. This package implements several ABC
  algorithms in a user friendly context to fit models in the spatial SEIRS class. 
  \tabular{ll}{
    Package: \tab ABSEIR\cr
    Type: \tab Package\cr
    Version: \tab 1.0\cr
    Date: \tab 2015-07-28\cr
    License: \tab GPL-3\cr}}
\references{
    Beaumont, M.A., Cornuet, J.-M., Marin, J.-M. et al., 2009. 
    Adaptive approximate Bayesian computation. Biometrika 96,983-990.

    Brown, Grant D., Oleson, Jacob J., Porter, Aaron T. 2015. An Empirically 
    Adjusted Approach to Reproductive Number Estimation for Stochastic 
    Compartmental Models: A Case Study of Two Ebola Outbreaks. Submitted to Biometrics

    Del Moral, P., Doucet, A., Jasra, A., 2012. An adaptive sequential Monte Carlo
    method for approximate Bayesian computation. Statistics and Computing 22, 1009-1020.

    Rubin, D., 1980. Bayesianly justifiable and relevant frequency calculations 
    for the applied statistician. The annals of Statistics 12, 1151-1172.         
}

\author{Grant Brown <grant-brown@uiowa.edu>}
\examples{
    print("See the vignettes")
}
