\name{ABSEIR-package}
\alias{ABSEIR-package}
\docType{package}
\title{Title goes here.}
\description{A description goes here.}
\details{
  A detailed description goes here. 
  \tabular{ll}{
    Package: \tab ABSEIR\cr
    Type: \tab Package\cr
    Version: \tab 0.01\cr
    Date: \tab 2015-07-28\cr
    License: \tab GPL-3\cr}}
\references{Manuscript in progress}
\author{Grant Brown <grant-brown@uiowa.edu>}
\examples{
    print("examples coming soon")
}
