% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataModel.R
\name{DataModel}
\alias{DataModel}
\title{Create a DataModel object, describing the relationship of your data to the SEIR model}
\usage{
DataModel(Y, type = c("identity", "overdispersion"),
  compartment = c("I_star", "R_star", "I"), cumulative = FALSE, phi = NA)
}
\arguments{
\item{Y}{a matrix with T rows and n columns, for time points and spatial locations respectively.}

\item{type}{a string equal to "identity" or "overdispersion"}

\item{compartment}{a string equal to "I_star" or "R_star"}

\item{cumulative}{a logical value indicating whether the data is reported on a
cumulative scale}

\item{phi}{optional overdispersion parameter}
}
\value{
an object of type \code{\link{DataModel}}
}
\description{
Create a DataModel object, describing the relationship of your data to the SEIR model
}
\details{
A fundamental task in building hierarchical models is to describe the way in which
 the observed data relates to the underlying model. Here, we assume that the
 matrix of observed values, \code{Y}, is related to one of two epidemic compartments:
 I_star or R_star. In the notation of Brown, Porter, and Oleson 2015, these correspond
 to the $I^*$ and $R^*$ compartments, which catalog the newly infectious and recovered
 individuals, respectively. For a discussion of these definitions, in addition to a more
 detailed description of the overall spatial SEIR framework, please refer to that work.
}
\examples{
dataModel = DataModel(rpois(100, 10))
}

