% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemicSimulations.R
\name{epidemic.simulations}
\alias{epidemic.simulations}
\title{perform and return epidemic simulations based on a fitted model object}
\usage{
epidemic.simulations(modelObject, replicates = 1, verbose = FALSE)
}
\arguments{
\item{modelObject}{a SpatialSEIRModel object, as created by the \code{\link{SpatialSEIRModel}}
function.}

\item{replicates}{the number of replicate simulations to perform for each sample
contained in \code{modelObject}
actual simulated compartment values should be returned. Omitting compartment
values is faster, but of course conveys less information.}

\item{verbose}{a logical value, indicating whether verbose output should be 
provided.}
}
\description{
perform and return epidemic simulations based on a fitted model object
}
\details{
The main SpatialSEIRModel functon performs many simulations, but for the sake of 
   memory efficiency and runtime does not return the simulated compartment values
   to the user. If simulated epidemics are desired, they may be quickly and
   easily generated using this function.
}
\examples{
\dontrun{simulate_values <- epidemic.simulations(modelObject, replicates = 10, 
                                                 verbose = TRUE)} 

}

