% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitionPriors.R
\name{PathSpecificTransitionPriors}
\alias{PathSpecificTransitionPriors}
\title{Build a path specific TransitionPriors object, which governs how individuals move from
the exposed to infectious and infectious to removed compartments.}
\usage{
PathSpecificTransitionPriors(Z1 = NA, Z2 = NA, truncation_prob = 1e-06)
}
\arguments{
\item{Z1}{a probability density function for the length of time individuals spend in 
the latent state}

\item{Z2}{a probability density function for the length of time individuals spend in 
the infectious state}

\item{truncation_prob}{ABSEIR precomputes the maximum number of discrete time points
during which a person may remain in the latent and infectious states. We therefore
require a probability cutoff. The trunction probability is the minimum that the
conditional probability of remaining in a disease state can become before individuals
are forced to transition.}
}
\description{
Build a path specific TransitionPriors object, which governs how individuals move from
the exposed to infectious and infectious to removed compartments.
}
\details{
The TransitionPriors component of spatial SEIR(S) models captures the  
 process by which individuals move from the exposed to infectious compartment, 
 and from the infectious to removed compartment. This component thus governs the
 duration of the latent and infectious periods of the disease of interest, on the
 discrete timescale employed.
}
\examples{
transitionPriors <- PathSpecificTransitionPriors(Z1 = function(x){dunif(x, 2, 10)},
                                        Z2 = function(x){dunif(x, 7, 24)})
}
\seealso{
\code{\link{TransitionPriors}}, \code{\link{ExponentialTransitionPriors}}
\code{\link{WeibullTransitionPriors}}
}
