% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transitionPriors.R
\name{TransitionPriors}
\alias{TransitionPriors}
\title{Build a TransitionPriors object, which governs how individuals move from
the exposed to infectious and infectious to removed compartments.}
\usage{
TransitionPriors(mode = c("exponential", "path_specific"), params = list())
}
\arguments{
\item{mode}{The type of transition model to employ}

\item{params}{Additional parameters, specific to the type of transition model. See
details section for additional information.}
}
\description{
Build a TransitionPriors object, which governs how individuals move from
the exposed to infectious and infectious to removed compartments.
}
\details{
The TransitionPriors component of spatial SEIR(S) models captures the
 process by which individuals move from the exposed to infectious compartment,
 and from the infectious to removed compartment. This component thus governs the
 duration of the latent and infectious periods of the disease of interest, on the
 discrete timescale employed.

 Two different TransitionPriors configurations are currently offered: the exponential
 compartment membership model, and a path specific generalization. These may be
 specified manually using the TransitionPriors function, or with the specific
 ExponentialTransitionPriors and PathSpecificTransitionPriors functions.

 The exponential version of this process requires four parameters:
 \itemize{
     \item{p_ei}{The probability, at a given time point, that an exposed individual will
                          become infectious}
     \item{p_ir}{The probability, at a given time point, that an infectious individual
                 be removed from the infectious population}
     \item{p_ei_ess}{An effective number of samples, corresponding to the confidence in
                     the chosen E to I transition probability}
     \item{p_ir_ess}{An effective number of samples, corresponding to the confidence in
                     the chosen I to R transition probability}
 }
 The path specific formulation requires fewer parameters, but more care is required in
 their specification.
 TODO: mention Z1, Z2
}
\examples{
transitionPriors <- TransitionPriors("exponential", params=list(p_ei=
                                        1/5, p_ir=1/7, p_ei_ess=100, p_ir_ess=100))
}

