% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_SpatialSEIRModel.R
\name{update.SpatialSEIRModel}
\alias{update.SpatialSEIRModel}
\title{Update a \code{\link{SpatialSEIRModel}} object by drawing additional samples.}
\usage{
\method{update}{SpatialSEIRModel}(object, ...)
}
\arguments{
\item{object}{a \code{\link{SpatialSEIRModel}} object}

\item{\dots}{Additional arguments include:
\itemize{
 \item{\code{sampling_control}: }{a \code{\link{SamplingControl}} object}
 \item{\code{samples}: }{indicates the number of parameter values to sample 
from the posterior}
 \item{\code{verbose}: }{indicates whether verbose output should be 
provided.}
}}
}
\description{
Update a \code{\link{SpatialSEIRModel}} object by drawing additional samples.
}
